% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finufft-package.R
\docType{package}
\name{finufft-package}
\alias{finufft-package}
\alias{_PACKAGE}
\title{finufft: R interface for finufft}
\description{
R interface to the FINUFFT library used to efficiently calculate nonuniform fast Fourier transforms.
}
