% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nfft}
\alias{nfft}
\title{nfft}
\usage{
nfft(x, c, N, tol = 1e-09)
}
\arguments{
\item{x}{locations}

\item{c}{complex weights}

\item{N}{number of output modes}

\item{tol}{precision}
}
\value{
complex vector
}
\description{
wrapper for 1d1 nufft
}
