% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nufft_1}
\alias{nufft_1}
\title{nufft_1}
\usage{
nufft_1(x, c, N1, tol = 1e-09, type = 1L)
}
\arguments{
\item{x}{locations}

\item{c}{complex weights}

\item{N1}{number of output modes}

\item{tol}{precision}

\item{type}{dimension}
}
\value{
complex vector
}
\description{
wrapper for 1d dim nufft
}
