% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nufft_1d1}
\alias{nufft_1d1}
\title{nufft_1d1}
\usage{
nufft_1d1(xj, cj, n1, tol = 1e-09)
}
\arguments{
\item{xj}{locations}

\item{cj}{complex weights}

\item{n1}{number of output modes}

\item{tol}{precision}
}
\value{
complex vector
}
\description{
wrapper for 1d dim nufft (dimension 1, type 1)
}
