% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nufft_1d2}
\alias{nufft_1d2}
\title{nufft_1d2}
\usage{
nufft_1d2(xj, fk, tol = 1e-09, iflag = 1L)
}
\arguments{
\item{xj}{locations}

\item{fk}{complex weights}

\item{tol}{precision}

\item{iflag}{+ or - i}
}
\value{
complex vector
}
\description{
wrapper for 1d dim nufft (dimension 1, type 2)
}
