% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nufft_1d3}
\alias{nufft_1d3}
\title{nufft_1d3}
\usage{
nufft_1d3(xj, cj, sk, tol = 1e-09)
}
\arguments{
\item{xj}{locations}

\item{cj}{complex weights}

\item{sk}{locations}

\item{tol}{precision}
}
\value{
complex vector
}
\description{
wrapper for 1d dim nufft (dimension 1, type 3)
}
