% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nufft_2}
\alias{nufft_2}
\title{nufft_2}
\usage{
nufft_2(x, y, c, N1, N2, tol = 1e-09, type = 1L)
}
\arguments{
\item{x}{locations}

\item{y}{locations}

\item{c}{complex weights}

\item{N1}{number of output modes}

\item{N2}{number of output modes}

\item{tol}{precision}

\item{type}{dimension}
}
\value{
complex vector
}
\description{
wrapper for 2d nufft
}
