% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nufft_2d1}
\alias{nufft_2d1}
\title{nufft_2d1}
\usage{
nufft_2d1(xj, yj, cj, n1, n2, tol = 1e-09, iflag = 1L)
}
\arguments{
\item{xj}{locations}

\item{yj}{locations}

\item{cj}{complex weights}

\item{n1}{number of output modes}

\item{n2}{number of output modes}

\item{tol}{precision}

\item{iflag}{+ or - i}
}
\value{
complex vector
}
\description{
wrapper for 2d dim nufft (dimension 2, type 1)
}
