% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nufft_2d2}
\alias{nufft_2d2}
\title{nufft_2d2}
\usage{
nufft_2d2(xj, yj, fk, tol = 1e-09, iflag = 1L)
}
\arguments{
\item{xj}{locations}

\item{yj}{locations}

\item{fk}{complex weights}

\item{tol}{precision}

\item{iflag}{+ or - i}
}
\value{
complex vector
}
\description{
wrapper for 2d dim nufft (dimension 2, type 2)
}
