% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nufft_2d3}
\alias{nufft_2d3}
\title{nufft_2d3}
\usage{
nufft_2d3(xj, yj, cj, sk, tk, tol = 1e-09, iflag = 1L)
}
\arguments{
\item{xj}{locations}

\item{yj}{locations}

\item{cj}{complex weights}

\item{sk}{locations}

\item{tk}{locations}

\item{tol}{precision}

\item{iflag}{+ or - i}
}
\value{
complex vector
}
\description{
wrapper for 2d dim nufft (dimension 2, type 3)
}
