% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nufft_3}
\alias{nufft_3}
\title{nufft_3}
\usage{
nufft_3(x, y, z, c, N1, N2, N3, tol = 1e-09, type = 1L)
}
\arguments{
\item{x}{locations}

\item{y}{locations}

\item{z}{locations}

\item{c}{complex weights}

\item{N1}{number of output modes}

\item{N2}{number of output modes}

\item{N3}{number of output modes}

\item{tol}{precision}

\item{type}{dimension}
}
\value{
complex vector
}
\description{
wrapper for 3d nufft
}
