% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nufft_3d1}
\alias{nufft_3d1}
\title{nufft_3d1}
\usage{
nufft_3d1(xj, yj, zj, cj, n1, n2, n3, tol = 1e-09, iflag = 1L)
}
\arguments{
\item{xj}{locations}

\item{yj}{locations}

\item{zj}{locations}

\item{cj}{complex weights}

\item{n1}{number of output modes}

\item{n2}{number of output modes}

\item{n3}{number of output modes}

\item{tol}{precision}

\item{iflag}{+ or - i}
}
\value{
complex vector
}
\description{
wrapper for 3d dim nufft (dimension 1, type 1)
}
