% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nufft_3d2}
\alias{nufft_3d2}
\title{nufft_3d2}
\usage{
nufft_3d2(xj, yj, zj, fk, tol = 1e-09, iflag = 1L)
}
\arguments{
\item{xj}{locations}

\item{yj}{locations}

\item{zj}{locations}

\item{fk}{complex weights}

\item{tol}{precision}

\item{iflag}{+ or - i}
}
\value{
complex vector
}
\description{
wrapper for 3d dim nufft (dimension 3, type 2)
}
