% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nufft_3d3}
\alias{nufft_3d3}
\title{nufft_3d3}
\usage{
nufft_3d3(xj, yj, zj, cj, sk, tk, uk, tol = 1e-09, iflag = 1L)
}
\arguments{
\item{xj}{locations}

\item{yj}{locations}

\item{zj}{locations}

\item{cj}{complex weights}

\item{sk}{locations}

\item{tk}{locations}

\item{uk}{locations}

\item{tol}{precision}

\item{iflag}{+ or - i}
}
\value{
complex vector
}
\description{
wrapper for 3d dim nufft (dimension 3, type 3)
}
